+++ changelog.md
@@ -1,7 +1,8 @@
 CHANGES:
 - Added T3 production technologies
-- Added ROI for rune technologium
+- Added ROI for rune technologium, mineral research centre and megalith
 
 FIXES:
 - Calculator settings button wordt nu niet meer dubbel aangemaakt.
 - De berekening om naar verzamelaar te gaan wordt nu ook goed uitgevoerd wanneer je meer processoren hebt dan het aantal dat gebruikt wordt.
+- De astro ROI bug gefixed waarbij